Fixed tests and applied some more refactoring
