--- src/CrossCutting.Utilities.ExpressionEvaluator.Sql/EvaluatableSqlExpressionProviderHandlers/EmptyEvaluatableHandler.cs
@@ -1,14 +1,0 @@
-namespace CrossCutting.Utilities.ExpressionEvaluator.Sql.EvaluatableSqlExpressionProviderHandlers;
-
-public class EmptyEvaluatableHandler : IEvaluatableSqlExpressionProviderHandler
-{
-    public async Task<Result<string>> GetExpressionAsync(object? context, IEvaluatable evaluatable, IFieldNameProvider fieldNameProvider, ParameterBag parameterBag, IEvaluatableSqlExpressionProviderHandler callback, CancellationToken token)
-    {
-        if (evaluatable is not EmptyEvaluatable && evaluatable is not EmptyEvaluatable<bool>)
-        {
-            return Result.Continue<string>();
-        }      
-        
-        return string.Empty;
-    }
-}
