+++ src/CrossCutting.Utilities.ExpressionEvaluator.Tests/EvaluatableSqlExpressionProviderHandlers/LiteralEvaluatableHandlerTests.cs
@@ -5,13 +5,13 @@ public class LiteralEvaluatableHandlerTests : TestBase<LiteralEvaluatableHandler
     public class GetExpressionAsync : LiteralEvaluatableHandlerTests
     {
         [Fact]
-        public async Task Returns_Correct_Sql()
+        public async Task Returns_Correct_Sql_For_Untyped_LiteralEvaluatable()
         {
             // Arrange
             var parameterBag = new ParameterBag();
             var evaluatable = new LiteralEvaluatableBuilder()
                 .WithValue("Hello world!")
-                .Build();
+                .BuildTyped();
             var sut = CreateSut();
             var callback = new EvaluatableSqlExpressionProvider([]);
 
@@ -23,6 +23,65 @@ public class LiteralEvaluatableHandlerTests : TestBase<LiteralEvaluatableHandler
             result.Value.ShouldBe("@p0");
             parameterBag.Parameters.Count.ShouldBe(1);
             parameterBag.Parameters.First().Value.ShouldBeEquivalentTo("Hello world!");
+        }
+    
+        [Fact]
+        public async Task Returns_Correct_Sql_For_Typed_LiteralEvaluatable()
+        {
+            // Arrange
+            var parameterBag = new ParameterBag();
+            var evaluatable = new LiteralEvaluatableBuilder<string>()
+                .WithValue("Hello world!")
+                .BuildTyped();
+            var sut = CreateSut();
+            var callback = new EvaluatableSqlExpressionProvider([]);
+
+            // Act
+            var result = await sut.GetExpressionAsync(null, evaluatable, Substitute.For<IFieldNameProvider>(), parameterBag, callback, CancellationToken.None);
+
+            // Assert
+            result.Status.ShouldBe(ResultStatus.Ok);
+            result.Value.ShouldBe("@p0");
+            parameterBag.Parameters.Count.ShouldBe(1);
+            parameterBag.Parameters.First().Value.ShouldBeEquivalentTo("Hello world!");
+        }
+
+        [Fact]
+        public async Task Returns_Correct_Sql_For_Untyped_EmptyEvaluatable()
+        {
+            // Arrange
+            var parameterBag = new ParameterBag();
+            var evaluatable = new EmptyEvaluatableBuilder().Build();
+            var sut = CreateSut();
+            var callback = new EvaluatableSqlExpressionProvider([]);
+
+            // Act
+            var result = await sut.GetExpressionAsync(null, evaluatable, Substitute.For<IFieldNameProvider>(), parameterBag, callback, CancellationToken.None);
+
+            // Assert
+            result.Status.ShouldBe(ResultStatus.Ok);
+            result.Value.ShouldBe("@p0");
+            parameterBag.Parameters.Count.ShouldBe(1);
+            parameterBag.Parameters.First().Value.ShouldBeNull();
+        }
+
+        [Fact]
+        public async Task Returns_Correct_Sql_For_Typed_EmptyEvaluatable()
+        {
+            // Arrange
+            var parameterBag = new ParameterBag();
+            var evaluatable = new EmptyEvaluatableBuilder<string>().Build();
+            var sut = CreateSut();
+            var callback = new EvaluatableSqlExpressionProvider([]);
+
+            // Act
+            var result = await sut.GetExpressionAsync(null, evaluatable, Substitute.For<IFieldNameProvider>(), parameterBag, callback, CancellationToken.None);
+
+            // Assert
+            result.Status.ShouldBe(ResultStatus.Ok);
+            result.Value.ShouldBe("@p0");
+            parameterBag.Parameters.Count.ShouldBe(1);
+            parameterBag.Parameters.First().Value.ShouldBeNull();
         }
     }
 }
